// packages/react/src/lib/components/avatar/avatar.d.ts
import * as React from 'react';
export declare const PdsAvatar: React.FunctionComponent<Partial<{
    variant: "default" | "color-bar";
    inverted: boolean;
    align: "default" | "center";
}> & {
    ref?: React.Ref<any> | undefined;
} & {
    style?: React.CSSProperties | undefined;
} & Partial<{}> & {
    id?: string | undefined;
    slot?: string | undefined;
    className?: string | undefined;
    children?: React.ReactNode;
}>;